pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teleclaude"
version = "0.1.0"
description = "Telegram bot framework with Claude Code session integration"
license = {text = "MIT"}
authors = [{name = "Ofir Cohen", email = "ofir5300@users.noreply.github.com"}]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["telegram", "claude", "bot", "ai", "claude-code"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/ofir5300/teleclaude"
Repository = "https://github.com/ofir5300/teleclaude"

[project.optional-dependencies]
dev = ["ruff", "pytest"]

[tool.setuptools.packages.find]
include = ["teleclaude*"]
